Technical Changes to Launcher Code
Switched from char arrays to Strings for simplicity of code.
It allows for easier comparisons of information that is posted to the
serial.
NOTE: It is understood that the terminator character (i.e. '\n' in this
case) is NOT included in the String that is returned from
"Serial.readStringUntil('\n')".