Updated Launcher and Hub Codes
Launcher code is complete.
Only thing left to do is further testing with new IR-Distance Sensor
(previous model did not meet specs).

Hub code was added to, significantly. However, there have been some
changes in the the methodology, which will result in a possible overhaul
of the process for the next update.